refactor: rewrite git and conventional commit parsing
In preparation for supporting tagging of go submodules, gotagger needs a
more flexible way to work with git repositories and commit messages.

This adds new internal git and commit packages. The new git package uses
go-git instead of subshells, and the new commit package aligns more
closely to the conventional commit standard. The original git and marker
packages remain, but are deprecated and will be removed before the
v1.0.0 release.

Change-Id: I063b47ff38e7f6fc58bf610ed99ea9b9b4121782