feat: refactor public Gotagger API
This change replaces the Gotagger `SubmoduleVersion` function with a
`ModuleVersions` function. The new `ModuleVersions` function takes a
variadic list of module names. When `ModuleVersions` is called with no
arguments, it reports the version of every go module in the repository,
but when called with one or more module names it only reports the
versions of those modules. `ModuleVersions` also preforms the same
release commit validation as `TagRepo`.

The Gotagger `Version` function is now explicitly for use in projects
that do not have any go modules or only one module.

Change-Id: Icd6cffa153f7bd19906b932c9bd3ba2f3afff3b8