fix: generate correct version for non-modules
The support for go modules broke versioning for projects that do not use
go modules anywhere, as the tag filtering excluded all tags with a major
version greater than 1. As part of fixing this bug, it was clear that
the go-git based code was part of the problem, so that code was replaced
with new calls to git itself.

Change-Id: Iff09826460d9702087a1d26615a9284d9558ce6e