feat: add option for handling dirty worktrees
In some development situations, users may want gotagger to report an
incremented version if the worktree is dirty, even if there are no
commits since the latest tag. This adds a new command-line option,
`-dirty`, and a Config member, `DirtyWorktreeIncrement`, that lets
users pick between a minor or patch increment.